+++ admin/src/pages/View/components/NavigationManager/index.tsx
@@ -113,33 +113,6 @@ export const NavigationManager = ({
   );
 };
 
-export const useNavigationManager = () => {
-  const [isOpened, setIsOpened] = useState(false);
-  const open = useCallback(() => setIsOpened(true), [setIsOpened]);
-  const close = useCallback(() => setIsOpened(false), [setIsOpened]);
-
-  const modal = useMemo(
-    () =>
-      isOpened ? (
-        <NavigationManager
-          initialState={{ view: "INITIAL" }}
-          isOpened
-          onClose={close}
-        />
-      ) : null,
-    [isOpened, close]
-  );
-
-  return useMemo(
-    () => ({
-      navigationManagerModal: modal,
-      openNavigationManagerModal: open,
-      closeNavigationManagerModal: close,
-    }),
-    [modal, open, close]
-  );
-};
-
 const renderHeader = (state: State) => {
   switch (state.view) {
     case "LIST":
