+++ admin/src/pages/View/index.js
@@ -28,7 +28,7 @@ import List from '../../components/NavigationItemList';
 import NavigationContentHeader from './components/NavigationContentHeader';
 import NavigationHeader from './components/NavigationHeader';
 import NavigationItemPopUp from "./components/NavigationItemPopup";
-import { useI18nCopyNavigationItemsModal } from "./components/I18nCopyNavigationItems";
+import { useI18nCopyNavigationItemsModal } from '../../hooks/useI18nCopyNavigationItemsModal';
 import Search from '../../components/Search';
 import useDataManager from "../../hooks/useDataManager";
 import { getTrad } from '../../translations';
@@ -38,7 +38,6 @@ import {
   usedContentTypes,
   validateNavigationStructure,
 } from './utils/parsers';
-import { useNavigationManager } from './components/NavigationManager';
 
 const View = () => {
   const {
